clean up

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: kgriset <kgriset@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/27 12:03:53 by kgriset           #+#    #+#             */
-/*   Updated: 2024/03/05 19:21:44 by kgriset          ###   ########.fr       */
+/*   Updated: 2024/03/05 20:02:27 by kgriset          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,7 +72,7 @@ typedef struct s_vars {
 
 // Render
 void render(t_vars *vars);
-int calc_mandelbrot(t_vars *vars);
+double calc_mandelbrot(t_vars *vars);
 
 // Events
 int key_events(int keycode, t_vars *vars);
